+++ libcloud/test/common/test_google.py
@@ -56,10 +56,13 @@ class GoogleBaseAuthConnectionTest(LibcloudTestCase):
     def setUp(self):
         GoogleBaseAuthConnection.conn_classes = (GoogleAuthMockHttp,
                                                  GoogleAuthMockHttp)
-        self.mock_scope = ['https://www.googleapis.com/auth/foo']
-        kwargs = {'scope': self.mock_scope}
+        self.mock_scopes = ['foo', 'bar']
+        kwargs = {'scopes': self.mock_scopes}
         self.conn = GoogleInstalledAppAuthConnection(*GCE_PARAMS,
                                                      **kwargs)
+
+    def test_scopes(self):
+        self.assertEqual(self.conn.scopes, 'foo bar')
 
     def test_add_default_headers(self):
         old_headers = {}
@@ -88,8 +91,8 @@ class GoogleInstalledAppAuthConnectionTest(LibcloudTestCase):
     def setUp(self):
         GoogleInstalledAppAuthConnection.conn_classes = (GoogleAuthMockHttp,
                                                          GoogleAuthMockHttp)
-        self.mock_scope = ['https://www.googleapis.com/auth/foo']
-        kwargs = {'scope': self.mock_scope}
+        self.mock_scopes = ['https://www.googleapis.com/auth/foo']
+        kwargs = {'scopes': self.mock_scopes}
         self.conn = GoogleInstalledAppAuthConnection(*GCE_PARAMS,
                                                      **kwargs)
 
@@ -128,15 +131,15 @@ class GoogleBaseConnectionTest(LibcloudTestCase):
     def setUp(self):
         GoogleBaseAuthConnection.conn_classes = (GoogleAuthMockHttp,
                                                  GoogleAuthMockHttp)
-        self.mock_scope = ['https://www.googleapis.com/auth/foo']
-        kwargs = {'scope': self.mock_scope, 'auth_type': 'IA'}
+        self.mock_scopes = ['https://www.googleapis.com/auth/foo']
+        kwargs = {'scopes': self.mock_scopes, 'auth_type': 'IA'}
         self.conn = GoogleBaseConnection(*GCE_PARAMS, **kwargs)
 
     def test_auth_type(self):
         self.assertRaises(GoogleAuthError, GoogleBaseConnection, *GCE_PARAMS,
                           **{'auth_type': 'XX'})
 
-        kwargs = {'scope': self.mock_scope}
+        kwargs = {'scopes': self.mock_scopes}
 
         if SHA256:
             kwargs['auth_type'] = 'SA'